+++ packages/app/rollup.chunks.js
@@ -97,15 +97,19 @@ const chunkTo = {     // Map of string -> (Regex | Array of Regex)
   //  /work/node_modules/@firebase/database/dist/index.esm2017.js
   //  /work/node_modules/@firebase/webchannel-wrapper/dist/index.esm2017.js
   //
-  //  /work/node_modules/idb/build/index.js   # used by Firebase app and performance
+  //  /work/node_modules/idb/build/index.js
   //
   "firebase-auth": /\/node_modules\/@firebase\/auth\//,
-  "firebase-database": /\/node_modules\/@firebase\/database\//,
-  "firebase-firestore": /\/node_modules\/@firebase\/firestore\//,
+  //"firebase-database": /\/node_modules\/@firebase\/database\//,
+  "firebase-firestore": [
+    /\/node_modules\/@firebase\/firestore\//,
+    /\/node_modules\/@firebase\/webchannel-wrapper\//,    // used by '@firebase/firestore'
+  ],
   "firebase-performance": /\/node_modules\/@firebase\/performance\//,
+  //"firebase-functions": /\/node_modules\/@firebase\/functions\//,   // just trying; used in the worker (NOT USED by the worker)
   "firebase": [
-    /\/node_modules\/@firebase\/(?:app|util|logger|component|installations|webchannel-wrapper)\//,
-    /\/node_modules\/idb\//,
+    /\/node_modules\/@firebase\/(?:app|util|logger|component|installations)\//,
+    /\/node_modules\/idb\//,      // needed by '@firebase/{app|installations|messaging}' (place in same chunk)
   ],
 
   // Plausible
@@ -142,68 +146,5 @@ const chunkTo = {     // Map of string -> (Regex | Array of Regex)
 
   // There should not be others. Production builds are banned with 'npm link'ed or 'file://') 'aside-keys'.
 };
-
-/*** OLD Vite 3.0.0-alpha.2 :
-const chunkTo = {     // Map of string -> (Regex | Array of Regex)
-
-  // default chunk; application itself and small stuff
-  "": [
-    /^\/work\/src\//,
-
-    // /work/prod/index.html
-    // /work/prod/index.html?html-proxy&index=0.js
-    // /work/prod/main.js
-    // /work/firebase.config.js
-    /^\/work\/prod\//,
-    /^\/work\/firebase\.config\.js$/,
-
-    // vite/preload-helper
-    // vite/modulepreload-polyfill
-    /^vite\/preload-helper$/,      // Vite runtime (small, ~600b)
-    /^vite\/modulepreload-polyfill$/,
-
-    // plugin-vue:export-helper
-    /^plugin-vue:export-helper/,  // very small, ~180b
-  ],
-
-  "vue": /\/node_modules\/@?vue\//,
-  "vue-router": /\/node_modules\/vue-router\//,
-  "aside-keys": /\/node_modules\/aside-keys\//,
-
-  // Firebase
-  //
-  // @firebase/{auth|firestore|app|util|logger|component|webchannel-wrapper|performance}
-  "firebase-auth": /\/node_modules\/@firebase\/auth\//,
-  "firebase": [
-    /\/node_modules\/@firebase\/(?:app|util|logger|component)\//,
-    /\/node_modules\/idb\//     // needed by '@firebase/{app|installations|messaging}' (place in the same chunk) (9.89k)
-  ],
-  "firebase-performance": [
-    /\/node_modules\/@firebase\/performance\//,
-    /\/node_modules\/@firebase\/installations\//,   // only used by '@firebase/performance' (so place in same chunk) (36.9k)
-  ],
-  "firebase-firestore": [
-    /\/node_modules\/@firebase\/firestore\//,
-    /\/node_modules\/@firebase\/webchannel-wrapper\//,
-  ],
-
-  // Sentry
-  //
-  // @sentry/{browser|tracing|core|utils|hub|minimal|types}
-  "sentry-browser": /\/node_modules\/@sentry\/browser\//,
-  "sentry-tracing": /\/node_modules\/@(sentry\/tracing)\//,
-
-  "sentry": /\/node_modules\/@sentry\/(?:core|utils|hub|minimal|types)\//,    // Note: 'minimal' is no more in Sentry 7.x
-
-  // TypeScript runtime
-  //
-  // NOTE: Seems important that this is placed in its own chunk (had it in the app chunk).
-  //
-  "tslib": /\/node_modules\/tslib\//,    // needed by Firebase and Sentry (15.2k)
-
-  // There should not be others. Production builds (where this code is involved) are banned with 'npm link'ed or
-  // 'file://') 'aside-keys'.
-};
-***/
 
 export { manualChunks };