+++ packages/app/tools/gen-vite-env-ci.js
@@ -3,7 +3,7 @@
 *
 * Usage:
 *   <<
-*     FIRESTORE_PORT=<num> AUTH_PORT=<num> \
+*     FIRESTORE_PORT=<num> AUTH_PORT=<num> FUNCTIONS_PORT=<num> \
 *     EMUL_HOST=<port> \
 *     [SENTRY_DSN=...] gen-vite-env-local --project=demo-...
 *   <<
@@ -45,12 +45,14 @@ const emulHost = process.env["EMUL_HOST"];
 
 const firestorePort = process.env["FIRESTORE_PORT"] || fail("Expected 'FIRESTORE_PORT' env.var.");
 const authPort = process.env["AUTH_PORT"] || fail("Expected 'AUTH_PORT' env.var.");
+const functionsPort = process.env["FUNCTIONS_PORT"] || fail("Expected 'FUNCTIONS_PORT' env.var.");
 
 const out =
 `# Generated
 #
 VITE_FIRESTORE_PORT=${firestorePort}
 VITE_AUTH_PORT=${authPort}
+VITE_FUNCTIONS_PORT=${functionsPort}
 VITE_PROJECT_ID=${projectId}
 VITE_EMUL_HOST=${emulHost}
 ${