+++ packages/backend/database.rules.js
@@ -1,8 +1,0 @@
-// Blocks all access; only used from admin (metrics and logging)
-
-export default {
-  "rules": {
-    ".read": false,
-    ".write": false
-  }
-}